Pixels are unsigned long.

+++ c/x_color.h
@@ -17,10 +17,10 @@ void col_close(Display *x_server, int x_screen, unsigned long color_id[], Colorm
 boolean col_check(int color_id);
 
 /* Gets the color_id of a color, non blinking on the background */
-int col_get_std(int background, int foreground, unsigned long color_id[]);
+unsigned long col_get_std(int background, int foreground, unsigned long color_id[]);
 
 /* Gets the color_id of a color, blinking on the background */
-int col_get_blk(int background, int foreground, unsigned long color_id[]);
+unsigned long col_get_blk(int background, int foreground, unsigned long color_id[]);
 
 /* Makes the colors blinking, or not blinking (alternatively) */
 boolean col_set_blinking (Display *x_server, int x_screen, unsigned long color_id[],
